Implement GhostInfo

+++ GbxTools3D/GbxTools3D.Client/Models/RenderDetails.cs
@@ -1,3 +1,3 @@
 ﻿namespace GbxTools3D.Client.Models;
 
-public sealed record RenderDetails(float? FPS, int? Calls, int? Triangles, int? Geometries, int? Textures);
+public sealed record RenderDetails(double? FPS, int? Calls, int? Triangles, int? Geometries, int? Textures);
